refactor(mero-abi): separate inspect and extract into modules
